Upgraded testing support to use greater than Bundler 0.9.19.
There were many api changes to the Bundler DSL between 0.8 and 0.9.
This included the removal of deprecated APIs, and a change in how
to load the bundler environment.

Now also instituting a new policy to always require a supporting
library in the lib/ ruby files even if they've already been
required. This allows the code reader to quickly understand what
DIRECT dependencies a file may have.